refactor(layout): split algorithm.ts into focused modules
Addresses the PR review (the file was too big). Pure move — output stays
bit-identical (layout + decoration goldens unchanged). algorithm.ts 1654 → 808
lines (the recursive packer core: _layoutDir, estimateDirReaches, layoutCity,
markJoinSides), with cohesive siblings:

- profiling.ts   — the opt-in per-phase profiler (was inline in the kernel)
- dimensions.ts  — file/building sizing + street-width derivation
- stemSolver.ts  — the forbidden-interval gap-fit packer (findSmallestValidStem,
                   placeChild, flips, trace types)
- overlaps.ts    — the runtime overlap diagnostic
- rect.ts        — gains the geometry primitives OVERLAP_EPS + _rectsOverlap

Importers (src + tests) repointed to the new modules.

Refs #63
