+++ app/tests/city/layout/layoutTrace.test.ts
@@ -3,8 +3,9 @@ import { WorldRectKind } from '@/city/layout/occupancyIndex';
 // findSmallestValidStem, placeChild, and the layoutCityWithTrace entry.
 
 import { describe, it, expect } from 'vitest';
-import { findSmallestValidStem, placeChild, layoutCityWithTrace } from '@/city/layout/algorithm';
-import type { VariantTrace } from '@/city/layout/algorithm';
+import { layoutCityWithTrace } from '@/city/layout/algorithm';
+import { findSmallestValidStem, placeChild } from '@/city/layout/stemSolver';
+import type { VariantTrace } from '@/city/layout/stemSolver';
 import { WorldOccupancy } from '@/city/layout/occupancyIndex';
 import { StreetAxis, NodeKind } from '@/types';
 import type { DirNode, FileNode, Manifest } from '@/types';
